✨ [Feature/auction-43] Comment 별점 조회/변경 - Comment 등록 시 별점 조회 및 멤버 별점 업데이트 진행 - 이후 경매 진행 이후 코멘트 등록이 가능하도록 유효성 검사 추가 예정 - 현재 AuctionCommentService Class에 있는 updateMemberGrade MemberService 클래스로 옮겨야 함

+++ src/main/java/com/selab/auction/participation/model/entity/AuctionSaleComment.java
@@ -34,10 +34,10 @@ public class AuctionSaleComment extends BaseEntity {
     private String comment;
 
     @Column
-    private Long grade;
+    private Double grade;
 
     @Builder
-    public AuctionSaleComment(long itemId, long buyMemberId, long saleMemberId, String comment, long grade){
+    public AuctionSaleComment(long itemId, long buyMemberId, long saleMemberId, String comment, double grade){
         this.itemId = itemId;
         this.buyMemberId = buyMemberId;
         this.saleMemberId = saleMemberId;
